cleanup kusama hotfix
